Move Home,Navbar & Quote components from components folder to pages folder and changed file extention
